Add files via upload
This code is a Compiler Design program that performs parenthesis checking. It is designed to validate a specific pattern in a given string. The code prompts the user to enter a string and checks if it satisfies the pattern requirements.

The program begins by initializing a boolean variable called "match" as false. It then prompts the user to enter a string and stores it in the variable "str".

Next, the code checks if the first and last characters of the string are both 'a'. If they are, it sets the "match" variable to true.

A while loop is used to iterate through the characters of the string, starting from the second character up to the second-to-last character. Within the loop, it checks if each character is equal to 'b'. If any character is found to be different from 'b', the "match" variable is set to false and the loop is exited using the break statement.

After the loop, the program checks the value of the "match" variable. If it is true, it outputs the message "Match found!" indicating that the string satisfies the pattern requirements. Otherwise, it outputs the message "No match found."

This code can be used as a reference or as a starting point for implementing more complex parenthesis checking algorithms in Compiler Design projects. It showcases the basics of string manipulation, conditional statements, and loops in the context of compiler-related tasks.